Show human readable time to serde repr
Adds a new field to the serde struct that represents a human readable
duration of time. This makes easier to debug the SCTE markers.

This commit also simplifies a bit the code using Default trait and other
simplifications.

Fixed the calculation of the sizes of sections in the final SCTE struct.
We were not calculating correctly the amount of bytes of the sections.
Now we use the Recorder pattern which returns the exact amount of bytes
written.